analyzer/requirements.txt
spacy
pdfminer.six
python-docx